% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{FilterDiagnosis2}
\alias{FilterDiagnosis2}
\title{Filter morbidity by secondary diagnosis}
\usage{
FilterDiagnosis2(data, diagnosis_id)
}
\arguments{
\item{data}{Morbidity data}

\item{diagnosis_id}{id of secondary diagnosis}
}
\value{
data frame with morbidity data prov_hosp, sexo, prov_res, diag_in, diag_ppal, motivo_alta, estancia, fecha_ingreso, edad
}
\description{
Filter morbidity by secondary diagnosis following international classification of diseases
}
\details{
Uses dplyr filter
}
\examples{
data <- data_ejemplo \%>\% FilterDiagnosis2(20)
}
