% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{FilterEmergency}
\alias{FilterEmergency}
\title{Filter morbidity by ER item}
\usage{
FilterEmergency(data)
}
\arguments{
\item{data}{Morbidity data}
}
\value{
data frame with morbidity data prov_hosp, sexo, prov_res, diag_in, diag_ppal, motivo_alta, estancia, fecha_ingreso, edad
}
\description{
Filter morbidity by ER
}
\details{
Uses dplyr filter
}
\examples{
data <- GetMorbiData(y1=2010,y2=2011) \%>\% FilterEmergency()
}
