% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{FilterProvincia}
\alias{FilterProvincia}
\title{Filter morbidity by provincia}
\usage{
FilterProvincia(data, provincia)
}
\arguments{
\item{data}{Morbidity data}

\item{provincia}{ID of provincia}
}
\value{
data frame with morbidity data prov_hosp, sexo, prov_res, diag_in, diag_ppal, motivo_alta, estancia, fecha_ingreso, edad
}
\description{
Filter morbidity by provincia using official id of provincias
}
\details{
Uses dplyr filter
}
\examples{
data <- GetMorbiData(y1=2010,y2=2011) \%>\% FilterProvincia(28)
}
