% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{GetMorbiData}
\alias{GetMorbiData}
\title{Download and read morbidity data for several years}
\usage{
GetMorbiData(y1 = 2005, y2 = 2015)
}
\arguments{
\item{year1}{Begining year of morbidity data}

\item{year2}{Ending year of morbidity data}
}
\value{
data frame with morbidity data prov_hosp, sexo, prov_res, diag_in, diag_ppal, motivo_alta, estancia, fecha_ingreso, edad
}
\description{
Download morbidity data from INE ftp and parse it
}
\details{
Uses ReadZip
}
\examples{
data <- GetMorbiData(y1=2010,y2=2011)
}
