% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{ReadZip}
\alias{ReadZip}
\title{Download and read morbidity data}
\usage{
ReadZip(year)
}
\arguments{
\item{year}{Year of morbidity data}
}
\value{
data frame with morbidity data prov_hosp, sexo, prov_res, diag_in, diag_ppal, motivo_alta, estancia, fecha_ingreso, edad
}
\description{
Download morbidity data from INE ftp and parse it
}
\details{
Uses read.fwf
}
\examples{
data <- ReadZip(2010)
}
