% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data.R
\name{SetPrevalence}
\alias{SetPrevalence}
\title{Calculate prevalence, relative values.}
\usage{
SetPrevalence(data, pop = "total")
}
\arguments{
\item{data}{Morbidity data}

\item{pop}{total, male, females}
}
\value{
relative values of prevalence
}
\description{
Calculate prevalence, relative values by total population of preovince
}
\details{
Uses poblacion
}
\examples{
data <- GetMorbiData(y1=2010,y2=2011) \%>\%  ReduceData(provincia = TRUE,date="year") \%>\% SetPrevalence()
}
