% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{source_lines}
\alias{source_lines}
\title{Auxiliary function to source code}
\usage{
source_lines(file, lines)
}
\arguments{
\item{year}{source file and lines}
}
\value{
execution of code
}
\description{
Source INE code, just the lines to maki it work
}
\details{
Uses script provided from INE
}
