% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{TraduceCIE10toCIE9}
\alias{TraduceCIE10toCIE9}
\title{Translate CIE-10 code to CIE-9}
\usage{
TraduceCIE10toCIE9(codigo)
}
\arguments{
\item{codigo}{code from morbidity data cie 10 version}
}
\value{
code from morbidity data cie 9 version
}
\description{
Get approximate CIE-9 code from CIE-10
}
\examples{
data <- TraduceCodigoEspecifico("S72109D")
}
