% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{TraduceCodigoEspecifico}
\alias{TraduceCodigoEspecifico}
\title{Get specific diagnosis}
\usage{
TraduceCodigoEspecifico(codigo)
}
\arguments{
\item{codigo}{code from morbidity data}
}
\value{
Specific diagnosis
}
\description{
Get specific diagnosis following international classification of diseases
}
\examples{
data <- TraduceCodigoEspecifico(3019)
}
