% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{AddDiagnosis2}
\alias{AddDiagnosis2}
\title{Add secondary diagnosis to morbity data}
\usage{
AddDiagnosis2(data)
}
\arguments{
\item{data}{Morbidity data}
}
\value{
data frame with morbidity data prov_hosp, sexo, prov_res, diag_in, diag_ppal, motivo_alta, estancia, fecha_ingreso, edad, diag2
}
\description{
Add secondary diagnosis following international classification of diseases
}
\examples{
data <- GetMorbiData(y1=2010,y2=2011) \%>\% AddDiagnosis2()
}
