% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data.R
\name{ReduceData}
\alias{ReduceData}
\title{Reduce data to temporal serie}
\usage{
ReduceData(data, provincia = TRUE, date = "day", diag = NULL,
  sex = FALSE)
}
\arguments{
\item{data}{Morbidity data}

\item{date}{day, month or year}

\item{diag}{diag1, daig2, diag3, if NULL not reduced by diagnosis}

\item{sex}{Boolean, reduce by sex}

\item{province}{Boolean, reduce by provicia}
}
\value{
data frame with data serie of morbidity
}
\description{
Reduce morbidity data to temporal serie, including diagnosis, province or sex
}
\details{
Uses group_by
}
\examples{
data <- GetMorbiData(y1=2010,y2=2011) \%>\% FilterProvincia(provincia = 28) \%>\% FilterEmergency() \%>\% AddDiagnosis1() \%>\% AddDiagnosis2() \%>\% FilterDiagnosis2(diagnosis_id = 19) \%>\% ReduceData(provincia = TRUE,date = "day")
}
